+++ pkg/lint/rulesets/rulesets.go
@@ -42,12 +42,16 @@ var ClusterApp = &RuleSet{
 		rules.AvoidUnevaluated{},
 		rules.PropertiesMustHaveOneType{},
 		rules.AvoidLogicalConstruct{},
+		rules.AdheresToCommonSchemaStructureRecommendations{},
+		rules.AdheresToCommonSchemaStructureRequirements{},
 		rules.ArrayItemsMustHaveSingleType{},
 		rules.AvoidRecursion{},
 		rules.AvoidRecursionKeywords{},
+		rules.ObjectsMustHaveProperties{},
 	},
 	excludeLocations: []string{
 		"/properties/internal",
+		"/properties/cluster-shared",
 	},
 }
 
@@ -93,14 +97,20 @@ func GetRuleSet(name RuleSetName) *RuleSet {
 	return ruleSet
 }
 
-func VerifyRuleSet(name string, schema *schemautils.ExtendedSchema) (errors []string, recommendations []string) {
+func VerifyRuleSet(
+	name string,
+	schema *schemautils.ExtendedSchema,
+) (errors []string, recommendations []string) {
 	nameEnum := RuleSetName(name)
 	ruleSet := GetRuleSet(nameEnum)
 
 	return LintWithRuleSet(schema, ruleSet)
 }
 
-func LintWithRuleSet(schema *schemautils.ExtendedSchema, ruleSet *RuleSet) (errors []string, recommendations []string) {
+func LintWithRuleSet(
+	schema *schemautils.ExtendedSchema,
+	ruleSet *RuleSet,
+) (errors []string, recommendations []string) {
 	errors = []string{}
 	recommendations = []string{}
 	for _, rule := range ruleSet.rules {